back/requirements.txt
fastapi
uvicorn[standard]
pymongo
pydantic
python-dotenv
gspread 
google-auth